requirements.txt
lpips==0.1.4
numpy==1.26.3
pillow==11.0.0
scikit-image==0.24.0
scikit-learn==1.6.1
pytorch-msssim==1.0.0
torch==2.6.0
torchvision==0.21.0
